+++ lib/src/badge_animation_type.dart
@@ -22,5 +22,5 @@ enum BadgeAnimationType {
 
   /// See also:
   /// * [RotationTransition]
-  rotation
+  rotation,
 }