+++ lib/src/painters/twitter_badge_shape_painter.dart
@@ -47,22 +47,38 @@ class TwitterBadgeShapePainter extends CustomPainter {
     }
 
     path.moveTo(width * 0.357, height * 0.156);
-    path.arcToPoint(Offset(width * 0.643, height * 0.156),
-        radius: Radius.circular(height * 0.157));
-    path.arcToPoint(Offset(width * 0.847, height * 0.396),
-        radius: Radius.circular(height * 0.165));
-    path.arcToPoint(Offset(width * 0.857, height * 0.666),
-        radius: Radius.circular(height * 0.170));
-    path.arcToPoint(Offset(width * 0.643, height * 0.844),
-        radius: Radius.circular(height * 0.163));
-    path.arcToPoint(Offset(width * 0.357, height * 0.844),
-        radius: Radius.circular(height * 0.157));
-    path.arcToPoint(Offset(width * 0.145, height * 0.665),
-        radius: Radius.circular(height * 0.163));
-    path.arcToPoint(Offset(width * 0.154, height * 0.372),
-        radius: Radius.circular(height * 0.170));
-    path.arcToPoint(Offset(width * 0.357, height * 0.156),
-        radius: Radius.circular(height * 0.163));
+    path.arcToPoint(
+      Offset(width * 0.643, height * 0.156),
+      radius: Radius.circular(height * 0.157),
+    );
+    path.arcToPoint(
+      Offset(width * 0.847, height * 0.396),
+      radius: Radius.circular(height * 0.165),
+    );
+    path.arcToPoint(
+      Offset(width * 0.857, height * 0.666),
+      radius: Radius.circular(height * 0.170),
+    );
+    path.arcToPoint(
+      Offset(width * 0.643, height * 0.844),
+      radius: Radius.circular(height * 0.163),
+    );
+    path.arcToPoint(
+      Offset(width * 0.357, height * 0.844),
+      radius: Radius.circular(height * 0.157),
+    );
+    path.arcToPoint(
+      Offset(width * 0.145, height * 0.665),
+      radius: Radius.circular(height * 0.163),
+    );
+    path.arcToPoint(
+      Offset(width * 0.154, height * 0.372),
+      radius: Radius.circular(height * 0.170),
+    );
+    path.arcToPoint(
+      Offset(width * 0.357, height * 0.156),
+      radius: Radius.circular(height * 0.163),
+    );
 
     paint.color = color!;
     canvas.drawPath(path, paint);