+++ lib/src/utils/gradient_utils.dart
@@ -2,6 +2,7 @@ import 'package:badges/badges.dart';
 import 'package:badges/src/badge_gradient_type.dart';
 import 'package:badges/src/utils/calculation_utils.dart';
 import 'package:material_ui/material_ui.dart';
+
 import 'dart:ui' as ui;
 
 class GradientUtils {