+++ test/badge_animations_tests/opacity_badge_animation_tests.dart
@@ -6,8 +6,9 @@ import '../test_widget_screen.dart';
 
 void opacityBadgeAnimationTests(badges.BadgeAnimationType badgeAnimationType) {
   if (badgeAnimationType != badges.BadgeAnimationType.fade) {
-    testWidgets('Opacity Badge With Appearance Animation Test',
-        (WidgetTester tester) async {
+    testWidgets('Opacity Badge With Appearance Animation Test', (
+      WidgetTester tester,
+    ) async {
       await tester.pumpWidget(
         TestWidgetScreen(
           toHideBadge: true,
@@ -59,8 +60,9 @@ void opacityBadgeAnimationTests(badges.BadgeAnimationType badgeAnimationType) {
 
   if (badgeAnimationType != badges.BadgeAnimationType.slide &&
       badgeAnimationType != badges.BadgeAnimationType.rotation) {
-    testWidgets('Opacity Badge Without Appearance Animation Test',
-        (WidgetTester tester) async {
+    testWidgets('Opacity Badge Without Appearance Animation Test', (
+      WidgetTester tester,
+    ) async {
       await tester.pumpWidget(
         TestWidgetScreen(
           toHideBadge: true,