+++ test/badge_animations_tests/show_hide_badge_animation_tests.dart
@@ -5,8 +5,9 @@ import 'package:flutter_test/flutter_test.dart';
 import '../test_widget_screen.dart';
 
 void showHideBadgeAnimationTests(badges.BadgeAnimationType badgeAnimationType) {
-  testWidgets('Show hide Badge Animation With Different Duration Test ',
-      (tester) async {
+  testWidgets('Show hide Badge Animation With Different Duration Test ', (
+    tester,
+  ) async {
     await tester.pumpWidget(
       TestWidgetScreen(
         badgeAnimationType: badgeAnimationType,
@@ -61,8 +62,9 @@ void showHideBadgeAnimationTests(badges.BadgeAnimationType badgeAnimationType) {
     expect(tester.hasRunningAnimations, false);
   });
 
-  testWidgets('Show hide Badge Animation Longer Appearance Duration Test ',
-      (tester) async {
+  testWidgets('Show hide Badge Animation Longer Appearance Duration Test ', (
+    tester,
+  ) async {
     await tester.pumpWidget(
       TestWidgetScreen(
         badgeAnimationType: badgeAnimationType,
@@ -114,8 +116,9 @@ void showHideBadgeAnimationTests(badges.BadgeAnimationType badgeAnimationType) {
     expect(tester.hasRunningAnimations, false);
   });
 
-  testWidgets('Show hide Badge Animation Without Appearance Duration Test ',
-      (tester) async {
+  testWidgets('Show hide Badge Animation Without Appearance Duration Test ', (
+    tester,
+  ) async {
     await tester.pumpWidget(
       TestWidgetScreen(
         badgeAnimationType: badgeAnimationType,
@@ -163,8 +166,9 @@ void showHideBadgeAnimationTests(badges.BadgeAnimationType badgeAnimationType) {
     expect(tester.hasRunningAnimations, false);
   });
 
-  testWidgets('Show hide Badge Animation Without Animation Duration Test ',
-      (tester) async {
+  testWidgets('Show hide Badge Animation Without Animation Duration Test ', (
+    tester,
+  ) async {
     await tester.pumpWidget(
       TestWidgetScreen(
         badgeAnimationType: badgeAnimationType,